styles MenuItemView and MenuItemForm components
