+++ src/components/MenuItemView/MenuItemView.module.css
@@ -1,0 +1,91 @@
+.item {
+    display: flex;
+    flex-wrap: wrap;
+    flex-direction: column;
+    padding: 1.5rem;
+    color: #211103;
+    background-color: #F9F8F8;
+    box-shadow: 2px 2px 2px #888888;
+    align-items: flex-start;
+    margin: .5rem;
+  }
+
+  .admin {
+      display: flex;
+      justify-content: flex-end;
+      align-items: center;
+  }
+
+  .admindel {
+    display: flex;
+    justify-content: flex-end;
+    align-items: center;
+    background-color: #F6F4D2;
+    padding: 1rem;
+    border: solid #E0E2DB 2px;
+    border-radius: 10px;  
+}
+
+.admindel p {
+  margin-right: 2rem;
+}
+
+
+  .admin div {
+      margin-left: .25rem;
+  }
+
+  .admin button {
+    border-radius: 10px;
+    font-size: 1rem;
+    padding: .5rem;
+    border: none;
+    color: #F9F8F8;
+    background-color: #5CAB7D;
+    box-shadow: 2px 2px 2px #888888;
+  }
+
+  .admin button:active {
+    background-color: green;
+    box-shadow: none;
+  }
+
+  .admindel button {
+    border-radius: 10px;
+    font-size: 1rem;
+    padding: .5rem;
+    border: none;
+    color: #F9F8F8;
+    background-color: #5CAB7D;
+    box-shadow: 2px 2px 2px #888888;
+  }
+
+  .admindel button:active {
+    background-color: green;
+    box-shadow: none;
+  }
+
+
+  .price {
+      width: 100%;
+  }
+
+  .price p {
+      font-weight: 900;
+  }
+
+.del button {
+  margin-left: 1rem;
+  border-radius: 10px;
+  font-size: 1rem;
+  padding: .5rem;
+  border: none;
+  color: #F9F8F8;
+  background-color: #A44A3F;
+  box-shadow: 2px 2px 2px #888888;
+}
+
+.del button:active {
+  background-color: #BF1A2F;
+  box-shadow: none;
+}
