+++ src/components/MenuItems/MenuItems.module.css
@@ -1,8 +1,10 @@
 .container {
+    /* border: red solid 2px; */
     display: flex;
     margin: 1.45rem;
     flex-direction: column;
     align-items: center;
+    justify-content: flex-start;
 }
 
 .item {
@@ -61,10 +63,11 @@
 }
 
 .category {
+    margin: 1rem;
     display: flex;
     flex-wrap: wrap;
-    align-items: center;
-    justify-content: center;
+    align-items: flex-start;
+    justify-content: space-around;
 }
 
 .del button {
