+++ src/pages/RestMenu/RestMenu.jsx
@@ -1,10 +1,11 @@
 import React from 'react';
 import { Redirect } from 'react-router-dom';
 import MenuItemForm from '../../components/MenuItemForm/MenuItemForm';
-import MenuItems from '../../components/MenuItem/MenuItems';
+import MenuItems from '../../components/MenuItems/MenuItems';
 import styles from './RestMenu.module.css';
 
 const Menu = (props) => {
+  
   return (
     <div>
       {props.myKitchen === null ? (
@@ -13,9 +14,11 @@ const Menu = (props) => {
         <div className={styles.container}>
           <div className={styles.header}>
             <h1>Menu Admin</h1>
-            <button id="addMenuItem" onClick={props.handleClick}>
-              Add New Item
-            </button>
+            { !props.menuItemForm ? 
+              <button id="addMenuItem" onClick={props.handleClick}>Add New Item</button>
+              :
+              null
+            }
           </div>
           <div className={styles.form}>
             {props.menuItemForm ? (
@@ -29,9 +32,12 @@ const Menu = (props) => {
           </div>
           <hr />
           <MenuItems
+            delMenu={props.delMenu}
             selectedMenuItem={props.selectedMenuItem}
             menuItems={props.menuItems}
             menuCats={props.menuCats}
+            handleDelMenu={props.handleDelMenu}
+            handleClick={props.handleClick}
             handleMenuItemEdit={props.handleMenuItemEdit}
             handleMenuItemUpdate={props.handleMenuItemUpdate}
             handleMenuItemDelete={props.handleMenuItemDelete}
